Rename make_functional_network_dicts.py to make_functional_network.py
